[DEV] Add utils.compare module (#226)
All comparison functions are included in
s/u/compare.py
The compareDocDecorator is included in s/u/docstrings.py

All are included in the __all__ lists at the top of the files
and included in the s/u/__init__.py

* DEV Add compareDictOfArrays comparison method

High-level method for comparing dictionaries of arrays.
Not intended to be a test method at this point, but heavily
relies upon the logging and other comparison methods.

The intent for this method is that this would be one of the
primary entry points for large comparisons.

* BUG Return boolean from ResultsReader.compareUniverses
* DEV Added some burnup values to results data w/o uncs

Added burnMaterials, burnMode, burnStep, iniBunFmass, totBurnFmass

* TST Add s/t/test_compare for testing comparison methods

Currently tests the compare method for the ResultsReader
with unmodified data and with modified data.
Utilized the TestCaseWithLogCapture to peek inside the
logs produced and ensure that proper outputs are present
